Nearly working
Have most code ready for performing analysis and ready for acceptance
tests.  While working on another project, noticed the output indexing in
original code is odd in that it takes the position after the maximum
instead of the maximum itself.  As testing that requires changes to code
from file * to char *, making commit now.  Also will address the last
position not printing in some cases.